requirements.txt
aiogram~=2.22.1
asyncpg~=0.26.0
emoji~=2.1.0
SQLAlchemy~=1.4.41
pyppeteer~=1.0.2
python-dotenv~=0.21.0